requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
agate==1.9.1
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
    #   dbt-snowflake
annotated-types==0.7.0
    # via pydantic
asn1crypto==1.5.1
    # via snowflake-connector-python
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
babel==2.18.0
    # via agate
boto3==1.42.90
    # via snowflake-connector-python
botocore==1.42.90
    # via
    #   boto3
    #   s3transfer
    #   snowflake-connector-python
build==1.4.3
    # via pip-tools
certifi==2025.1.31
    # via
    #   dbt-snowflake
    #   requests
    #   snowflake-connector-python
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via
    #   requests
    #   snowflake-connector-python
click==8.3.2
    # via
    #   -r requirements.in
    #   dbt-core
    #   dbt-semantic-interfaces
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   dbt-common
cryptography==46.0.7
    # via
    #   pyopenssl
    #   snowflake-connector-python
daff==1.4.2
    # via dbt-core
dbt-adapters==1.22.10
    # via
    #   dbt-core
    #   dbt-snowflake
dbt-common==1.37.3
    # via
    #   dbt-adapters
    #   dbt-core
    #   dbt-snowflake
dbt-core==1.11.8
    # via dbt-snowflake
dbt-extractor==0.6.0
    # via dbt-core
dbt-protos==1.0.443
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
dbt-semantic-interfaces==0.9.0
    # via dbt-core
dbt-snowflake==1.11.4
    # via -r requirements.in
deepdiff==8.6.2
    # via dbt-common
dotenv==0.9.9
    # via -r requirements.in
filelock==3.28.0
    # via snowflake-connector-python
idna==3.11
    # via
    #   requests
    #   snowflake-connector-python
importlib-metadata==8.9.0
    # via dbt-semantic-interfaces
isodate==0.7.2
    # via
    #   agate
    #   dbt-common
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.1.2
    # via keyring
jaraco-functools==4.4.0
    # via keyring
jinja2==3.1.6
    # via
    #   dbt-common
    #   dbt-core
    #   dbt-semantic-interfaces
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
jsonschema==4.26.0
    # via
    #   dbt-common
    #   dbt-core
    #   dbt-semantic-interfaces
jsonschema-specifications==2025.9.1
    # via jsonschema
keyring==25.7.0
    # via snowflake-connector-python
leather==0.4.1
    # via agate
markupsafe==3.0.3
    # via jinja2
mashumaro[msgpack]==3.14
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
more-itertools==10.8.0
    # via
    #   dbt-semantic-interfaces
    #   jaraco-classes
    #   jaraco-functools
msgpack==1.1.2
    # via mashumaro
nba-api==1.11.4
    # via -r requirements.in
networkx==3.6.1
    # via dbt-core
numpy==2.4.4
    # via
    #   nba-api
    #   pandas
orderly-set==5.5.0
    # via deepdiff
packaging==26.1
    # via
    #   build
    #   dbt-core
    #   snowflake-connector-python
    #   wheel
pandas==2.3.3
    # via
    #   nba-api
    #   snowflake-connector-python
parsedatetime==2.6
    # via agate
pathspec==0.12.1
    # via
    #   dbt-common
    #   dbt-core
pip-tools==7.5.3
    # via -r requirements.in
platformdirs==4.9.6
    # via snowflake-connector-python
protobuf==6.33.6
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
    #   dbt-protos
pyarrow==23.0.1
    # via snowflake-connector-python
pycparser==3.0
    # via cffi
pydantic==2.13.2
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
pydantic-core==2.46.2
    # via pydantic
pyjwt==2.12.1
    # via snowflake-connector-python
pyopenssl==26.0.0
    # via snowflake-connector-python
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   dbt-common
    #   dbt-semantic-interfaces
    #   pandas
python-dotenv==1.2.2
    # via
    #   -r requirements.in
    #   dotenv
python-slugify==8.0.4
    # via agate
pytimeparse==1.1.8
    # via agate
pytz==2026.1.post1
    # via
    #   dbt-adapters
    #   dbt-core
    #   pandas
    #   snowflake-connector-python
pywin32-ctypes==0.2.3
    # via keyring
pyyaml==6.0.3
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.33.1
    # via
    #   dbt-common
    #   dbt-core
    #   nba-api
    #   snowflake-connector-python
    #   snowplow-tracker
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
snowflake-connector-python[pandas,secure-local-storage]==4.4.0
    # via
    #   -r requirements.in
    #   dbt-snowflake
snowplow-tracker==1.1.0
    # via dbt-core
sortedcontainers==2.4.0
    # via snowflake-connector-python
sqlparse==0.5.5
    # via dbt-core
text-unidecode==1.3
    # via python-slugify
tomlkit==0.14.0
    # via snowflake-connector-python
typing-extensions==4.15.0
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
    #   dbt-semantic-interfaces
    #   mashumaro
    #   pydantic
    #   pydantic-core
    #   pyopenssl
    #   referencing
    #   snowflake-connector-python
    #   snowplow-tracker
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2026.1
    # via
    #   agate
    #   pandas
urllib3==2.6.3
    # via
    #   botocore
    #   requests
wheel==0.46.3
    # via pip-tools
zipp==3.23.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
